+++ CHANGELOG.md
@@ -4,9 +4,17 @@
 
 ### Added
 
+- example script to export entire EGM data (Robert Arnold)
+- import/export of additional beat annotations in reference channel (Robert Arnold)
+- export of PaSo reference template for CARTO studies (Robert Arnold)
+
 ### Changed
 
+- made function call unprotected: CartoPoint.import_attributes() (Robert Arnold)
+
 ### Fixed
+
+- error when creating nested output folders (Robert Arnold)
 
 ### Removed